Added Chrome Extension
Since chrome extensions are one of the subject matters covered in most (all?) programming classes here I thought it would be beneficial to have a section for chrome extensions.

The link I added is to the getting started guide which explains all the parts of a chrome extension.  This will help students get started on making chrome extensions.