Added Udacity course on Android Programming
I chose this because at a glance, it looks like the mini projects that we get assigned. I think that this course would be useful because a lot of people rely on apps for a lot of different things, so being able to make an app is an important skill to have if you work with technology and computers.