Added link to Vettery
Vettery is a website where you make a profile and companies reach out to you and offer you jobs. This is important for students because jobs at this age give you valuable job experience which is useful for the rest of your life. It also gives you an idea of what you can do as a career. This method also makes the process of applying less stressful for students who already have a lot of stress in their life.